All applications are combined into one 'Debt'
