--- apps/clients/admin.py
@@ -1,10 +1,0 @@
-from django.contrib import admin
-
-from.models import Client
-
-
-@admin.register(Client)
-class ClientAdmin(admin.ModelAdmin):
-    list_display = ("id", "last_name", "first_name", "patronymic", "ipn")
-    list_filter = ("last_name", "first_name", "patronymic", "ipn")
-    search_fields = ("last_name__startswith", "first_name__startswith", "patronymic__startswith", "ipn__startswith")
