--- apps/contractors/choices.py
@@ -1,8 +1,0 @@
-TYPE_CONTRACTOR = [
-    ('МФО', 'Мікрофінансова організація'),
-    ('ФК', 'Факторингова компанія'),
-    ('БУ', 'Банкова установа'),
-    ('СК', 'Страхова компанія'),
-    ('ТК', 'Телекомунікаційна компанія'),
-    ('ЖКГ', 'Житлово-комунальне гопподарство')
-]
