+++ apps/debts/choices.py
@@ -3,3 +3,12 @@ CURRENCIES = [
     ('EUR', 'Euro'),
     ('UAH', 'Hryvnia'),
 ]
+
+TYPE_CONTRACTOR = [
+    ('МФО', 'Мікрофінансова організація'),
+    ('ФК', 'Факторингова компанія'),
+    ('БУ', 'Банкова установа'),
+    ('СК', 'Страхова компанія'),
+    ('ТК', 'Телекомунікаційна компанія'),
+    ('ЖКГ', 'Житлово-комунальне гопподарство')
+]
