+++ settings/settings.py
@@ -33,9 +33,7 @@ ALLOWED_HOSTS = ['127.0.0.1', '.pythonanywhere.com']
 
 ROOT_APPS = [
     'apps.core',
-    'apps.clients',
     'apps.debts',
-    'apps.contractors',
 ]
 
 INSTALLED_APPS = [
